+++ example.py
@@ -32,23 +32,25 @@ def create_config():
         config.option_from_alias('string', name='STRING_FROM_ALIAS', default='Default string value'),
         ConfigOption(
             name='OS',
-            option_type=ConfigOptionType.STRING,
-            default=platform.system(),
-            external=True
+            option_type=ConfigOptionType.EXTERNAL,
+            default=platform.system()
+        ),
+        ConfigOption(
+            name='OS_LIVE',
+            option_type=ConfigOptionType.EXTERNAL,
+            default=lambda: platform.system()
         ),
         ConfigOption(
             name='PYTHON_EVALUATED',
-            option_type=ConfigOptionType.STRING,
-            default=".".join(map(str, sys.version_info[:3])),
-            dependencies=lambda x: x.ENABLE_FEATURE_A,
-            external=True
+            option_type=ConfigOptionType.EXTERNAL,
+            default=".".join(map(str, sys.version_info[:3]))
         ),
     )
 
     ### Actions can also be added using a decorator for ease
     @config.action_option(
         requires=lambda x: x.LOG_LEVEL, 
-        dependencies="ENABLE_FEATURE_A",
+        dependencies=lambda x: x.ENABLE_FEATURE_A,
     )
     def build(x):
         print("Building...")
@@ -56,7 +58,7 @@ def create_config():
         return True
     
     # First define a group
-    deployment_group_proxy = config.group_option("deployment", dependencies="ENABLE_FEATURE_A")
+    deployment_group_proxy = config.group_option("deployment", dependencies=lambda x: x.ENABLE_FEATURE_A)
 
     # The group ConfigOption can be accessed using the get() function and
     # sub options can also be added to the group by directly appending them
@@ -71,7 +73,7 @@ def create_config():
     # Then use the group's action_option decorator to add actions easier
     @deployment_group_proxy.action_option(
         requires=lambda x: x.build(),
-        dependencies="ENABLE_FEATURE_A"
+        dependencies=lambda x: x.ENABLE_FEATURE_A
     )
     def deploy(x):
         print("Deploying...")