+++ CHANGELOG.md
@@ -27,6 +27,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ---
 
+## [0.12.1]
+### Added
+- Added missing power operator
+- New syntax for simplicity
+
+### Changed
+- Syntax of the json file
+
+### Fixed
+- Power operator
+- Operator order
+- Fixed search for groups
+
+### Deprecated
+- Old json syntax
+
+---
+
 ## [0.10.6]
 ### Added
 - Added support for multiple config files